chatbot_dev/requirements.txt
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
llama-cpp-python==0.2.16
faiss-cpu==1.7.4
pymongo==4.5.0
python-dotenv==1.0.0
numpy==1.26.0
tqdm==4.66.1
